[riscv] moving vla meta-function from vla_vectorizer to vla_function module
